Crate libits

Generic IoT3 messaging and telemetry for V2X

MQTT

Generic MQTT exchanges can be managed using both MqttClient to publish and subscribe and MqttRouter to handle messages using different callbacks depending the Topic a message is received on.

If the telemetry feature is enabled:

message reception and publishing will be traced1

the W3C Context will be injected as an MQTTv5 user property before publishing a message

Telemetry

Available with the telemetry feature

Provides functions to send traces to an OpenTelemetry collector

V2X

Available with the mobility feature

Provides traits and functions to create applications that will listen to MQTT messages

Geo routing

Available with the geo_routing feature

Provides topic management using queue system and quadkey positioning

For now, two traces are sent, but the aim is to send a single trace carrying (at least) two spans: one for reception and one for publishing ↩

Modules

client

Client implementation and configuration

transport

MQTT and Telemetry clients implementation

Functions

now

Returns the current UTC timestamp in milliseconds