Function libits::client::bootstrap::bootstrap

pub async fn bootstrap(ini: Ini) -> Result<Configuration, ConfigurationError>

Calls the bootstrap API and builds a Configuration out of bootstrap information

Bootstrap sequence will return the URL and credentials to use to connect to the services (MQTT, OTLP collector, …), any of these information already present in the configuration file would be overridden
All the other fields that are unrelated to services connection are still read from the configuration file and set in the Configuration object

Example of the bootstrap configuration section:

[bootstrap]
host="mydomain.com"
port=1234
path="/bootstrap"
role="external-app"
username="boot"
password="str4P!"